backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
redis==4.6.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
requests==2.31.0
groq>=0.3.0
python-dotenv==1.0.0
python-multipart==0.0.6